combined AutoDiffRev object into auto_diff.py
